rename: image 위치 변경(common 디렉터리 -> 상위 디렉터티)
