+++ src/main/java/labweb/labweb_be/image/domain/ImageType.java
